python/pyproject.toml
[project]
name = "ol-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openlineage-python>=1.30.1",
]
